src/quantum_consciousness/core/quantum_gns_construction.py
"""
Quantum GNS (Gelfand-Naimark-Segal) construction implementation.
Implements geometric and topological aspects of quantum systems.
"""

import numpy as np
from typing import List, Tuple, Dict, Optional, Callable
from scipy.linalg import expm, sqrtm, logm
from dataclasses import dataclass

@dataclass
class GNSState:
    """Represents a GNS construction state."""
    vector: np.ndarray
    algebra_element: np.ndarray
    representation: np.ndarray

class QuantumGNSConstruction:
    """Implements quantum GNS construction."""

    def __init__(self, dim: int):
        """
        Initialize GNS construction.

        Args:
            dim: Dimension of quantum system
        """
        self.dim = dim
        self.epsilon = 1e-12
        self.hilbert_space = self._initialize_hilbert_space()

    def _initialize_hilbert_space(self) -> np.ndarray:
        """Initialize Hilbert space basis."""
        return np.eye(self.dim)

    def construct_state(self, functional: Callable[[np.ndarray], complex]) -> GNSState:
        """
        Construct GNS state from functional
        π(a)Ω = |ψ_a⟩

        Args:
            functional: State functional

        Returns:
            GNSState: Constructed GNS state
        """
        # Create cyclic vector
        omega = np.ones(self.dim) / np.sqrt(self.dim)

        # Create algebra element
        algebra = np.zeros((self.dim, self.dim), dtype=complex)
        for i in range(self.dim):
            for j in range(self.dim):
                basis = np.zeros((self.dim, self.dim))
                basis[i,j] = 1
                algebra[i,j] = functional(basis)

        # Create representation
        representation = self._create_representation(algebra)

        return GNSState(
            vector=omega,
            algebra_element=algebra,
            representation=representation
        )

    def _create_representation(self, algebra_element: np.ndarray) -> np.ndarray:
        """Create GNS representation."""
        # Ensure algebra element is Hermitian
        H = algebra_element + algebra_element.conj().T
        # Create unitary representation
        return expm(-1j * H)

    def inner_product(self, state1: GNSState, state2: GNSState) -> complex:
        """
        Calculate GNS inner product
        ⟨ψ_a|ψ_b⟩ = ω(a*b)

        Args:
            state1: First GNS state
            state2: Second GNS state

        Returns:
            complex: Inner product value
        """
        product = state1.algebra_element.conj().T @ state2.algebra_element
        return np.vdot(state1.vector, product @ state2.vector)

    def geometric_phase(self, state: GNSState, parameter: float) -> float:
        """
        Calculate geometric phase
        γ = i∫⟨ψ|d/dt|ψ⟩dt

        Args:
            state: GNS state
            parameter: Evolution parameter

        Returns:
            float: Geometric phase
        """
        # Create parameter-dependent state
        evolved_state = expm(-1j * parameter * state.algebra_element) @ state.vector

        # Calculate Berry connection
        connection = -1j * np.vdot(state.vector, evolved_state)

        return np.angle(connection)

    def construct_modular_flow(self, state: GNSState, time: float) -> np.ndarray:
        """
        Construct modular flow
        σ_t(x) = Δ^{it}xΔ^{-it}

        Args:
            state: GNS state
            time: Flow parameter

        Returns:
            numpy.ndarray: Modular flow operator
        """
        # Calculate modular operator
        S = state.algebra_element @ state.representation
        delta = S.conj().T @ S

        # Calculate modular flow
        return expm(1j * time * np.log(delta + self.epsilon * np.eye(self.dim)))

    def tomita_takesaki_modular(self, state: GNSState) -> Tuple[np.ndarray, np.ndarray]:
        """
        Implement Tomita-Takesaki modular theory
        S = J∆^{1/2}

        Args:
            state: GNS state

        Returns:
            Tuple[np.ndarray, np.ndarray]: Modular operator and conjugation
        """
        # Calculate polar decomposition
        S = state.algebra_element @ state.representation
        delta = S.conj().T @ S
        delta_sqrt = sqrtm(delta + self.epsilon * np.eye(self.dim))

        # Calculate modular conjugation
        J = S @ np.linalg.inv(delta_sqrt + self.epsilon * np.eye(self.dim))

        return delta, J

    def relative_entropy(self, state1: GNSState, state2: GNSState) -> float:
        """
        Calculate relative modular entropy
        S(ω_1|ω_2) = -⟨ξ_1|log(∆_2)|ξ_1⟩

        Args:
            state1: First GNS state
            state2: Second GNS state

        Returns:
            float: Relative entropy
        """
        # Calculate relative modular operator
        S1 = state1.algebra_element @ state1.representation
        S2 = state2.algebra_element @ state2.representation
        delta_rel = np.linalg.inv(S1.conj().T @ S1 + self.epsilon * np.eye(self.dim)) @ \
                   (S2.conj().T @ S2 + self.epsilon * np.eye(self.dim))

        # Calculate relative entropy
        log_delta = logm(delta_rel + self.epsilon * np.eye(self.dim))
        return -np.real(np.vdot(state1.vector, log_delta @ state1.vector))

    def construct_crossed_product(self, state: GNSState,
                                action: np.ndarray) -> np.ndarray:
        """
        Construct crossed product algebra
        M ⋊_α G

        Args:
            state: GNS state
            action: Group action

        Returns:
            numpy.ndarray: Crossed product representation
        """
        # Create crossed product representation
        dim_total = self.dim * len(action)
        crossed = np.zeros((dim_total, dim_total), dtype=complex)

        for i, g in enumerate(action):
            block = state.representation @ g
            crossed[i*self.dim:(i+1)*self.dim, i*self.dim:(i+1)*self.dim] = block

        return crossed

    def compute_index(self, state: GNSState) -> float:
        """
        Compute Jones index
        [M:N] = dim(L²(M))/dim(L²(N))

        Args:
            state: GNS state

        Returns:
            float: Jones index
        """
        # Calculate dimensions using traces
        trace_M = np.abs(np.trace(state.representation @ state.representation.conj().T))
        trace_N = np.abs(np.trace(state.algebra_element @ state.algebra_element.conj().T))

        return trace_M / (trace_N + self.epsilon)

    def construct_tower(self, state: GNSState, depth: int) -> List[np.ndarray]:
        """
        Construct Jones tower
        N ⊂ M ⊂ M₁ ⊂ M₂ ⊂ ...

        Args:
            state: Initial GNS state
            depth: Tower depth

        Returns:
            List[np.ndarray]: Jones tower algebras
        """
        tower = [state.algebra_element]
        current = state.representation

        for _ in range(depth):
            # Basic construction step
            next_level = np.kron(current, current.conj().T)
            tower.append(next_level)
            current = next_level

        return tower

    def compute_categorical_trace(self, morphism: np.ndarray) -> complex:
        """
        Compute categorical trace
        tr(f) = ∑ᵢ⟨eᵢ|f|eᵢ⟩

        Args:
            morphism: Morphism matrix

        Returns:
            complex: Categorical trace
        """
        basis = self.hilbert_space
        trace = 0j

        for i in range(self.dim):
            trace += np.vdot(basis[i], morphism @ basis[i])

        return trace
